Clean up op program benchmark
